split header
